adding databases and classes

+++ KKAP Stars/Assets/Contestant.cs
@@ -2,14 +2,14 @@
 {
     private string name;
     private string company;
-    private double singing;
-    private double rapping;
-    private double dancing;
-    private double likability;
+    private int singing;
+    private int rapping;
+    private int dancing;
+    private int likability;
     private double starPower;
 
-    public Contestant(string newName, string newCompany, double newSinging, double newRapping, double newDancing,
-        double newLikability, double newStarPower)
+    public Contestant(string newName, string newCompany, int newSinging, int newRapping, int newDancing,
+        int newLikability, double newStarPower)
     {
         name = newName;
         company = newCompany;
@@ -40,42 +40,42 @@
         return company;
     }
 
-    public void setSinging(double newSinging)
+    public void setSinging(int newSinging)
     {
         singing = newSinging;
     }
 
-    public double getSinging()
+    public int getSinging()
     {
         return singing;
     }
 
-    public void setRapping(double newRapping)
+    public void setRapping(int newRapping)
     {
         rapping = newRapping;
     }
 
-    public double getRapping()
+    public int getRapping()
     {
         return rapping;
     }
 
-    public void setDancing(double newDancing)
+    public void setDancing(int newDancing)
     {
         dancing = newDancing;
     }
 
-    public double getDancing()
+    public int getDancing()
     {
         return dancing;
     }
 
-    public void setLikability(double newLikability)
+    public void setLikability(int newLikability)
     {
         likability = newLikability;
     }
 
-    public double getLikability()
+    public int getLikability()
     {
         return likability;
     }
